Refactor LegoOmni into LegoMain/Scripts
