+++ ISLE/isleapp.cpp
@@ -6,8 +6,8 @@
 #include "legobuildingmanager.h"
 #include "legogamestate.h"
 #include "legoinputmanager.h"
+#include "legomain.h"
 #include "legomodelpresenter.h"
-#include "legoomni.h"
 #include "legopartpresenter.h"
 #include "legovideomanager.h"
 #include "legoworldpresenter.h"
