+++ LEGO1/lego/legoomni/src/actors/bike.cpp
@@ -3,10 +3,10 @@
 #include "isle_actions.h"
 #include "legocontrolmanager.h"
 #include "legogamestate.h"
-#include "legoomni.h"
 #include "legoutils.h"
 #include "legoworld.h"
 #include "misc.h"
+#include "scripts.h"
 
 DECOMP_SIZE_ASSERT(Bike, 0x164)
