+++ LEGO1/lego/legoomni/src/common/misc.cpp
@@ -2,11 +2,12 @@
 
 #include "3dmanager/lego3dmanager.h"
 #include "legoinputmanager.h"
-#include "legoomni.h"
+#include "legomain.h"
 #include "legovideomanager.h"
 #include "mxbackgroundaudiomanager.h"
 #include "mxmisc.h"
 #include "roi/legoroi.h"
+#include "scripts.h"
 
 // GLOBAL: LEGO1 0x100f4c58
 MxBool g_isWorldActive = TRUE;
