+++ LEGO1/lego/legoomni/src/entity/legoworld.cpp
@@ -11,8 +11,8 @@
 #include "legogamestate.h"
 #include "legoinputmanager.h"
 #include "legolocomotionanimpresenter.h"
+#include "legomain.h"
 #include "legonavcontroller.h"
-#include "legoomni.h"
 #include "legoplantmanager.h"
 #include "legosoundmanager.h"
 #include "legoutils.h"
