+++ LEGO1/lego/legoomni/src/entity/legoworldpresenter.cpp
@@ -5,8 +5,8 @@
 #include "legoanimationmanager.h"
 #include "legobuildingmanager.h"
 #include "legoentity.h"
+#include "legomain.h"
 #include "legomodelpresenter.h"
-#include "legoomni.h"
 #include "legopartpresenter.h"
 #include "legoplantmanager.h"
 #include "legotexturepresenter.h"
