+++ LEGO1/lego/legoomni/src/gasstation/gasstationentity.cpp
@@ -6,11 +6,11 @@
 #include "islepathactor.h"
 #include "legoanimationmanager.h"
 #include "legogamestate.h"
-#include "legoomni.h"
 #include "legoutils.h"
 #include "legoworld.h"
 #include "misc.h"
 #include "mxtransitionmanager.h"
+#include "scripts.h"
 
 DECOMP_SIZE_ASSERT(GasStationEntity, 0x68)
