+++ LEGO1/lego/legoomni/src/video/legovideomanager.cpp
@@ -2,7 +2,7 @@
 
 #include "3dmanager/lego3dmanager.h"
 #include "legoinputmanager.h"
-#include "legoomni.h"
+#include "legomain.h"
 #include "misc.h"
 #include "mxdirectx/mxdirect3d.h"
 #include "mxdirectx/mxstopwatch.h"
