+++ LEGO1/lego/legoomni/src/worlds/elevatorbottom.cpp
@@ -7,7 +7,7 @@
 #include "legocontrolmanager.h"
 #include "legogamestate.h"
 #include "legoinputmanager.h"
-#include "legoomni.h"
+#include "legomain.h"
 #include "legovariables.h"
 #include "misc.h"
 #include "mxmisc.h"
