+++ LEGO1/lego/legoomni/src/worlds/gasstation.cpp
@@ -8,7 +8,7 @@
 #include "legocontrolmanager.h"
 #include "legogamestate.h"
 #include "legoinputmanager.h"
-#include "legoomni.h"
+#include "legomain.h"
 #include "misc.h"
 #include "mxbackgroundaudiomanager.h"
 #include "mxmisc.h"
@@ -18,6 +18,7 @@
 #include "mxtimer.h"
 #include "mxtransitionmanager.h"
 #include "radio.h"
+#include "scripts.h"
 
 DECOMP_SIZE_ASSERT(GasStation, 0x128)
