+++ LEGO1/lego/legoomni/src/worlds/jukebox.cpp
@@ -7,7 +7,7 @@
 #include "legocontrolmanager.h"
 #include "legogamestate.h"
 #include "legoinputmanager.h"
-#include "legoomni.h"
+#include "legomain.h"
 #include "misc.h"
 #include "mxmisc.h"
 #include "mxnotificationmanager.h"
