+++ LEGO1/lego/legoomni/src/worlds/registrationbook.cpp
@@ -5,7 +5,6 @@
 #include "legocontrolmanager.h"
 #include "legogamestate.h"
 #include "legoinputmanager.h"
-#include "legoomni.h"
 #include "misc.h"
 #include "mxactionnotificationparam.h"
 #include "mxbackgroundaudiomanager.h"
@@ -17,6 +16,7 @@
 #include "mxtimer.h"
 #include "mxtransitionmanager.h"
 #include "regbook_actions.h"
+#include "scripts.h"
 
 DECOMP_SIZE_ASSERT(RegistrationBook, 0x2d0)
